pyproject.toml
[tool.poetry]
name = "first-discord-bot"
version = "0.1.0"
description = ""
authors = ["Little-Wretch <nazarr914@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
discord = "^2.3.2"
logging = "^0.4.9.6"
youtube-dl = "^2021.12.17"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
